Odds and Evens from Microsoft: DEV276x (Java)
This project focused on the use of while loops or if-else statements. It is a game called "Odds and Evens", similar to rock paper scissors. It is played between two players, the user versus the computer. Each player picks either "odds" or "evens", the user gets the first pick. Then, each player chooses a number of fingers to play, from 0 to 5. The winner is determined by whether the sum of all fingers is odd or even (depending on what players chose).